requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
cryptography>=41.0.0
streamlit>=1.30.0
